progress toward real multicenter
The content of etics::scf::scfclass::CalculateGravity has be set up like that it only calls other member functions; this way an equivalent routine in the multicenter module can imitate it.

N removed from the cache structure (the cache is initialized with Nmax), it should be passed as a parameter for the kernels that use it. In fact, maybe it's better to pass the pointer to the cache as a kernel parameter instead of sticking it in global memory. This could help later on when the particles feeling the force are different from the particles from which the force is calculated.